some fix
If serialize returns an empty string, make sure the id attributes
include an underscore. They must be in the form: "set_number" For
example, a 3 element list with id attributes foo_1, foo_5, foo_2 will
serialize to foo[]=1&foo[]=5&foo[]=2. You can use an underscore, equal
sign or hyphen to separate the set and number. For example foo=1 or
foo-1 or foo_1 all serialize to foo[]=1.